+++ frontend/src/App.css
@@ -38,7 +38,7 @@ a{
 
 /* feed */
 .feedContainer { 
-  padding-top: 3%; 
+  padding-top: 4%; 
   width: 100%; 
 }
