+++ frontend/src/components/Search/SearchBar.js
@@ -1,9 +1,10 @@
 import React, { Component } from 'react';
 import { slide as Menu } from 'react-burger-menu'; 
+import { Link } from "react-router-dom";
 import './SearchBar.css'
-import cookbooklogo from './cookbooknamelogo.png'
-import writingicon from './writingicon.png'
-import hearticon from './hearticon.png'
+import cookbooklogo from '../../images/cookbooknamelogo.png'
+import writingicon from '../../images/writingicon.png'
+import hearticon from '../../images/hearticon.png'
 
 class Searchbar extends Component {
   constructor() {
@@ -24,10 +25,14 @@ class Searchbar extends Component {
       <img className="searchbarLogoName" src={cookbooklogo} />
       <img className="searchbarLogo" src="http://irfanyurdu.org/wp-content/uploads/2017/04/eat-flat-1.png" />
       <input className="searchInput" type="search" name="search" placeholder="Search for recipes or users"/> 
+      <Link to={`/addrecipe`} className="searchLink">
       <img src={writingicon} className="writingIcon"/>
       <p className="addTagline"> Add recipe </p> 
+      </Link>
+      <Link to={`/favorites`} className="searchLink">
       <img src={hearticon} className="heartIcon"/>
       <p className="heartTagline"> Favorite recipes </p> 
+      </Link> 
       <div>
       <Menu right className="burgerMenu">
       <a id="home" className="menu-item" href="/">Home</a>
